# frontend/schema/openapi.json
{
  "components": {
    "schemas": {
      "ActionResponse": {
        "properties": {
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "state": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GameStateResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "title": "ActionResponse",
        "type": "object"
      },
...
      "GameStateResponse": {
        "properties": {
          "config": {
            "additionalProperties": true,
            "title": "Config",
            "type": "object"
          },
          "current_set": {
            "title": "Current Set",
            "type": "integer"
          },
          "match_finished": {
            "title": "Match Finished",
            "type": "boolean"
          },
          "serve": {
            "title": "Serve",
            "type": "string"
          },
          "simple_mode": {
            "title": "Simple Mode",
            "type": "boolean"
          },
          "team_1": {
            "$ref": "#/components/schemas/TeamState"
          },
          "team_2": {
            "$ref": "#/components/schemas/TeamState"
          },
          "visible": {
            "title": "Visible",
            "type": "boolean"
          }
        },
        "required": [
          "current_set",
          "visible",
          "simple_mode",
          "match_finished",
          "team_1",
          "team_2",
          "serve",
          "config"
        ],
        "title": "GameStateResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ServeRequest": {
        "properties": {
          "team": {
            "enum": [
              1,
              2
            ],
            "title": "Team",
            "type": "integer"
          }
        },
        "required": [
          "team"
        ],
        "title": "ServeRequest",
        "type": "object"
      },
      "SetScoreRequest": {
        "properties": {
          "set_number": {
            "minimum": 1.0,
            "title": "Set Number",
            "type": "integer"
          },
          "team": {
            "enum": [
              1,
              2
            ],
            "title": "Team",
            "type": "integer"
          },
          "value": {
            "maximum": 99.0,
            "minimum": 0.0,
            "title": "Value",
            "type": "integer"
          }
        },
        "required": [
          "team",
          "set_number",
          "value"
        ],
        "title": "SetScoreRequest",
        "type": "object"
      },
      "SetSetsRequest": {
        "properties": {
          "team": {
            "enum": [
              1,
              2
            ],
            "title": "Team",
            "type": "integer"
          },
          "value": {
            "maximum": 9.0,
            "minimum": 0.0,
            "title": "Value",
            "type": "integer"
          }
        },
        "required": [
          "team",
          "value"
        ],
        "title": "SetSetsRequest",
        "type": "object"
      },
...
      "TeamActionRequest": {
        "properties": {
          "team": {
            "enum": [
              1,
              2
            ],
            "title": "Team",
            "type": "integer"
          },
          "undo": {
            "default": false,
            "title": "Undo",
            "type": "boolean"
          }
        },
        "required": [
          "team"
        ],
        "title": "TeamActionRequest",
        "type": "object"
      },
      "TeamState": {
        "properties": {
          "scores": {
            "additionalProperties": true,
            "title": "Scores",
            "type": "object"
          },
          "serving": {
            "title": "Serving",
            "type": "boolean"
          },
          "sets": {
            "title": "Sets",
            "type": "integer"
          },
          "timeouts": {
            "title": "Timeouts",
            "type": "integer"
          }
        },
        "required": [
          "sets",
          "timeouts",
          "scores",
          "serving"
        ],
        "title": "TeamState",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Volley Overlay Control",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/customization": {
      "get": {
        "operationId": "get_customization_api_v1_customization_get",
        "parameters": [
          {
            "description": "Overlay ID",
            "in": "query",
            "name": "oid",
            "required": true,
            "schema": {
              "description": "Overlay ID",
              "title": "Oid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Customization",
        "tags": [
          "Scoreboard API v1"
        ]
      },
      "put": {
        "operationId": "update_customization_api_v1_customization_put",
        "parameters": [
          {
            "description": "Overlay ID",
            "in": "query",
            "name": "oid",
            "required": true,
            "schema": {
              "description": "Overlay ID",
              "title": "Oid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Data",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Customization",
        "tags": [
          "Scoreboard API v1"
        ]
      }
    },
...
    "/api/v1/game/add-point": {
      "post": {
        "operationId": "add_point_api_v1_game_add_point_post",
        "parameters": [
          {
            "description": "Overlay ID",
            "in": "query",
            "name": "oid",
            "required": true,
            "schema": {
              "description": "Overlay ID",
              "title": "Oid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Point",
        "tags": [
          "Scoreboard API v1"
        ]
      }
    },
    "/api/v1/game/add-set": {
      "post": {
        "operationId": "add_set_api_v1_game_add_set_post",
        "parameters": [
          {
            "description": "Overlay ID",
            "in": "query",
            "name": "oid",
            "required": true,
            "schema": {
              "description": "Overlay ID",
              "title": "Oid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Set",
        "tags": [
          "Scoreboard API v1"
        ]
      }
    },
    "/api/v1/game/add-timeout": {
      "post": {
        "operationId": "add_timeout_api_v1_game_add_timeout_post",
        "parameters": [
          {
            "description": "Overlay ID",
            "in": "query",
            "name": "oid",
            "required": true,
            "schema": {
              "description": "Overlay ID",
              "title": "Oid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Timeout",
        "tags": [
          "Scoreboard API v1"
        ]
      }
    },
    "/api/v1/game/change-serve": {
      "post": {
        "operationId": "change_serve_api_v1_game_change_serve_post",
        "parameters": [
          {
            "description": "Overlay ID",
            "in": "query",
            "name": "oid",
            "required": true,
            "schema": {
              "description": "Overlay ID",
              "title": "Oid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Change Serve",
        "tags": [
          "Scoreboard API v1"
        ]
      }
    },
    "/api/v1/game/reset": {
      "post": {
        "operationId": "reset_game_api_v1_game_reset_post",
        "parameters": [
          {
            "description": "Overlay ID",
            "in": "query",
            "name": "oid",
            "required": true,
            "schema": {
              "description": "Overlay ID",
              "title": "Oid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reset Game",
        "tags": [
          "Scoreboard API v1"
        ]
      }
    },
    "/api/v1/game/set-score": {
      "post": {
        "operationId": "set_score_api_v1_game_set_score_post",
        "parameters": [
          {
            "description": "Overlay ID",
            "in": "query",
            "name": "oid",
            "required": true,
            "schema": {
              "description": "Overlay ID",
              "title": "Oid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetScoreRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Score",
        "tags": [
          "Scoreboard API v1"
        ]
      }
    },
    "/api/v1/game/set-sets": {
      "post": {
        "operationId": "set_sets_api_v1_game_set_sets_post",
        "parameters": [
          {
            "description": "Overlay ID",
            "in": "query",
            "name": "oid",
            "required": true,
            "schema": {
              "description": "Overlay ID",
              "title": "Oid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetSetsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Sets",
        "tags": [
          "Scoreboard API v1"
        ]
      }
    },
...
  }
}